feat: add role based managment to services

+++ api/app/Http/Controllers/ServiceController.php
@@ -4,12 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Models\Service;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ServiceController extends Controller
 {
     //
     public function index()
     {
+        Gate::authorize('manage-services');
+
         return Service::all();
     }
 
@@ -20,6 +23,8 @@ class ServiceController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize('manage-services');
+
         $request->validate(['name' => 'required']);
         $service = Service::create($request->only('name'));
 
@@ -28,6 +33,8 @@ class ServiceController extends Controller
 
     public function update(Request $request, Service $service)
     {
+        Gate::authorize('manage-services');
+
         $request->validate(['name' => 'required']);
         $service->update($request->only('name'));
 
@@ -36,6 +43,8 @@ class ServiceController extends Controller
 
     public function destroy(Service $service)
     {
+        Gate::authorize('manage-services');
+
         $service->delete();
 
         return response()->json(null, 204);
